+++ src/main/java/dev/hoyin1600p/vhaccelerator/backport/BackportFeature.java
@@ -98,7 +98,7 @@ public enum BackportFeature {
             "Forge tag-registry concurrency",
             BackportSide.COMMON,
             false,
-            false,
+            true,
             List.of(
                     "bugfix.concurrency.MappedRegistryMixin",
                     "bugfix.concurrency.NamespacedWrapperMixin",